+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added Playwright testing library (DR-3680)
+- Added Playwright tests for navigation menu, what is public domain and search bar for the home page (DR-3658)
 
 ## [0.4.2] 2025-05-22